modified database, service, frontend

+++ frontend/src/app/services/agency.service.ts
@@ -2,17 +2,20 @@ import { Injectable } from '@angular/core';
 import { CONFIG } from '../common/config';
 import { Agency } from '../models/agency';
 import { WebRequestService } from './web-request.service';
+import { Helper } from '../helpers/helper';
 
 @Injectable({ providedIn: 'root' })
 export class AgencyService {
     readonly url: string = CONFIG.URL.AGENCY;
+    readonly helper = new Helper();
 
     constructor(
         private webrequestService: WebRequestService,
     ) { }
 
     getAgencyList() {
-        return this.webrequestService.get(this.url);
+        const agencyId = this.helper.getAgencyId();
+        return this.webrequestService.get(this.url + `/${agencyId}`);
     }
 
     create(obj: Agency) {
@@ -34,25 +37,19 @@ export class AgencyService {
     update(obj: Agency) {
         const payload = {
             id: obj.id,
+            userId: obj.userId,
             agencyName: obj.agencyName,
             address: obj.address,
             phone: obj.phone,
             note: obj.note,
             email: obj.email,
-            userName: obj.userName,
-            password: obj.password,
             contract: obj.contract,
-            role: obj.role,
             updatedByUserId: obj.updatedByUserId,
         };
         return this.webrequestService.put(this.url, payload);
     }
 
-    delete(id: number) {
-        return this.webrequestService.delete(this.url + `/${id}`);
-    }
-
-    changePasswordAdmin(password: string) {
-        return this.webrequestService.post(CONFIG.URL.USER + '/change/password/admin', { password });
+    delete(id: number, userId: number) {
+        return this.webrequestService.delete(this.url + `/${id}/${userId}`);
     }
 }
